requirements.txt
langchain
langchain-community
langchain-core
langchain_groq
chromadb
fpdf
SpeechRecognition
pyttsx3
transformers
torch
pyaudio
unstructured
sentence-transformers
load_dotenv
langchain_huggingface
